row kernel and stream_compact optimization row: 37.4ms -> 7.74ms (4.8x) getGaussian() -> Float32 by default, blocksize optimization for occupancy only problems left in ncu are GMEM access patterns (13.29%) and non-fused fp32(5.87%) stream_compact: 2406.40us -> 407.84us (5.9x) type specified all variables to 4 bytes blocksize optimization for occupancy bitwise operators for binary arithmetic demoted gmem and shmem count type rem and div in place of % and ÷ (no promotion)

+++ header.jl
@@ -1,10 +1,10 @@
 struct potential_blob
-	thisX::Int64 # 1
-	y::Int32 # 2
-	thisImg::Int32 # 3
-	x::Int64 # 4
-	oct::Int32 # 5
-	lay::Int32 # 6
+	thisX::UInt32 # 1
+	y::UInt32 # 2
+	thisImg::UInt32 # 3
+	x::UInt32 # 4
+	oct::UInt32 # 5
+	lay::UInt32 # 6
 end
 
 potential_blob() = potential_blob(0, 0, 0, 0, 0, 0)
